feat: Create a GroupChannelListItem and View component (#869)
[CLNP-1594](https://sendbird.atlassian.net/browse/CLNP-1594)

### Fix
* Create a `GroupChannelListItem` component instead of the
`ChannelPreview` component
* Create a `GroupChannelListItemView` component for the presentation
a layer of those components
* Apply the presentation layer component to the both components
