+++ src/modules/GroupChannelList/components/GroupChannelListItem/channel-preview.scss
