+++ src/modules/GroupChannelList/components/GroupChannelListItem/utils.ts
